requirements.txt
click~=7.1.2
numpy>=1.19.1
pillow~=7.2.0
tqdm~=4.49.0
opencv-python>=4.3.0.36
detectron2~=0.2
setuptools>=49.2.1
pycocotools>=2.0.1
torchvision>=0.5.0